fix(review): move BuildActivityFeedInput to interfaces, harden imports
Move BuildActivityFeedInput into activity-feed.interface.ts alongside
ActivityFeedItem, per the shared-package convention that types live in
interfaces/ and values live in constants/ (self-serve-code-reviewer).
Also switch activity-feed.utils.ts's '../interfaces' import to `import
type`, so it's erased entirely and can't reintroduce the barrel-import
cycle even if a future interface file adds a runtime dependency
(general-code-reviewer). Narrows the votingEnabled doc comment to
scope its claim to the activity feed only — the pre-existing pending
actions / active-votes stat paths read votes() independently and are
unaffected by this flag; addressing those is out of this ticket's scope.

LFXV2-1716
