+++ packages/shared/src/interfaces/activity-feed.interface.ts
@@ -1,6 +1,11 @@
 // Copyright The Linux Foundation and each contributor to LFX.
 // SPDX-License-Identifier: MIT
 
+import type { CommitteeDocument } from './committee.interface';
+import type { PastMeeting } from './meeting.interface';
+import type { Vote } from './poll.interface';
+import type { Survey } from './survey.interface';
+
 /**
  * Source kind backing an `ActivityFeedItem`.
  * @description Drives icon + tab-navigation choice in the group Overview activity feed stop-gap.
@@ -27,4 +32,21 @@ export interface ActivityFeedItem {
   icon: string;
   /** Tab-navigation context string passed to the existing `tab:context` handler, e.g. "meetings:upcoming" */
   tab: string;
+}
+
+export interface BuildActivityFeedInput {
+  pastMeetings: PastMeeting[];
+  votes: Vote[];
+  surveys: Survey[];
+  documents: CommitteeDocument[];
+  /**
+   * Vote items are excluded entirely when false. The Overview activity feed navigates a clicked vote
+   * row to the Votes tab, but that tab is hidden when the committee has voting disabled, and
+   * tab-navigation only validates against the static valid-tabs list — not tab visibility — so an
+   * activity row for a vote from before voting was disabled would otherwise navigate to a hidden,
+   * blank tab. This only closes that gap for the activity feed itself; the committee-overview
+   * "My Pending Actions" / "Active Votes" surfaces read `votes()` independently and are unaffected by
+   * this flag — see committee-overview.component.ts.
+   */
+  votingEnabled: boolean;
 }
